#195
git-svn-id: svn://projects.jkraemer.net/acts_as_ferret/trunk@303 7326d000-0a0e-0410-9cad-a9b28e7838dc

+++ plugin/acts_as_ferret/lib/unix_daemon.rb
@@ -49,6 +49,7 @@ module ActsAsFerret
       #################################################################################
       # create the PID file and install an at_exit handler
       def write_pid_file
+        raise "ferret_server may already be running, a pid file exists: #{@cfg.pid_file}" if read_pid_file
         open(@cfg.pid_file, "w") {|f| f << Process.pid << "\n"}
         at_exit { File.unlink(@cfg.pid_file) if read_pid_file == Process.pid }
       end
